requirements.txt
setuptools
logzero
git+https://github.com/Rapptz/discord.py
asyncpg
python-Levenshtein
tabulate~=0.8.7
pyyaml~=5.3.1
pylast==3.1.0
qrcode==6.1
pillow
aiohttp~=3.6.2
git+https://github.com/Rapptz/discord-ext-menus
webcolors~=1.3
jishaku
discord-flags
toml